+++ Samples/auth/Office-Add-in-NodeJS-SSO/Complete/app.js
@@ -13,8 +13,6 @@ var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
-var getGraphData = require('./public/javascripts/server-msgraph-helper');
-
 var indexRouter = require('./routes/index');
 var getFilesRoute = require('./routes/getFilesRoute');
 
@@ -52,37 +50,8 @@ app.get('/dialog.html', (async (req, res) => {
   return res.sendfile('dialog.html');
 }));
 
-app.get('/getuserdata', getFilesRoute);
+app.get('/getuserfilenames', getFilesRoute);
 
-// app.get('/getuserdata', async function(req, res, next) {
-//   const graphToken = req.get('access_token');
-
-//   // Minimize the data that must come from MS Graph by specifying only the property we need ("name")
-//   // and only the top 10 folder or file names.
-//   // Note that the last parameter, for queryParamsSegment, is hardcoded. If you reuse this code in
-//   // a production add-in and any part of queryParamsSegment comes from user input, be sure that it is
-//   // sanitized so that it cannot be used in a Response header injection attack. 
-//   const graphData = await getGraphData(graphToken, "/me/drive/root/children", "?$select=name&$top=10");
-
-//   // If Microsoft Graph returns an error, such as invalid or expired token,
-//   // there will be a code property in the returned object set to a HTTP status (e.g. 401).
-//   // Relay it to the client. It will caught in the fail callback of `makeGraphApiCall`.
-//   if (graphData.code) {
-//       next(createError(graphData.code, "Microsoft Graph error " + JSON.stringify(graphData)));
-//   }
-//   else 
-//   {
-//     // MS Graph data includes OData metadata and eTags that we don't need.
-//     // Send only what is actually needed to the client: the item names.
-//     const itemNames = [];
-//     const oneDriveItems = graphData['value'];
-//     for (let item of oneDriveItems){
-//         itemNames.push(item['name']);
-//     }
-
-//     res.send(itemNames)
-//   }
-// });
 
 
 // Catch 404 and forward to error handler
