+++ Samples/auth/Office-Add-in-NodeJS-SSO/Complete/public/javascripts/data.js
@@ -6,7 +6,7 @@
 
 function makeGraphApiCall(accessToken) {
     $.ajax({type: "GET", 
-        url: "/getuserdata",
+        url: "/getuserfilenames",
         headers: {"access_token": accessToken },
         cache: false
     }).done(function (response) {
@@ -22,7 +22,7 @@ function makeGraphApiCall(accessToken) {
         });
     })
     .fail(function (errorResult) {
-        // This error is relayed from `app.get('/getuserdata` in app.js file.
+        // This error is relayed from `app.get('/getuserfilenames` in app.js file.
         showMessage("Error from Microsoft Graph: " + JSON.stringify(errorResult));
 	});
 }
