+++ Samples/auth/Office-Add-in-NodeJS-SSO/Complete/server-helpers/obo-auth-helper.js
