+++ Samples/auth/Office-Add-in-NodeJS-SSO/Complete/server-helpers/odata-helper.js
